requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg==3.1.14
python-dotenv==1.0.0